+++ README.md
@@ -16,11 +16,3 @@
   - Run dev server with `npm run start:dev` (dev server with watch mode)
   - Build produciton code with `npm run build`
   - Start prod server with `npm run`
-
-
-## TODO
-  - endpoints to implement
-    - /login    <- default route
-    - /logout
-    - /register
-    - /         <- some simple hello word response on authenticated user
